Add futures
Futures wraps incomplete computations, so the program
can continue its usual flow, and wait for those computations
until it really needs the results.

The idea is to use them to fetch slow results (like DB access,
or a web service), wrapping those access in a Future, and passing
the future to the templates (allowing better parallelization).

Signed-off-by: Esteban RB <estebarb@gmail.com>