Major Ion refactorization (WIP)
I wasn't liking the way in which the context
handling was evolving at Ion: is verbose, is fragile,
is ugly.

The previous attempts used:
- A global map[request](map[string]interface{})
- The same datastructure as above, but per component

The first one was less ugly and less verbose, but it
was fragile.

The second one was an attempt to make the whole context
thing more robust, but it only got ugly and verbose.

So, in this attempt I'm experimenting with the idea
of a single value in the request context: one value
passed by the program.

In this case the program would have to create a struct
with all the fields that it needs to work (or that the
components used needs to work), and this struct would
automatically implement the interfaces provided by
the components to access its data.

So, it is a little bit more verbose than the first attempt,
a lot less than the secondm but more important: is more
robust. It is using the type system to enforce what is
in the context and what not.

Also, this new system works well with the new context
in http.Request, and also still works, without changing
any line of code, with previous Go versions.

This commit contains part of the work made to do this
changes, but not all. In particular the template
package is completelly broken, but I want to experiment
a bit with this branch before deciding which would be
a better way to pass the context to the templates.

Signed-off-by: Esteban Rodríguez Betancourt <estebarb@gmail.com>